@@ -1,6 +1,8 @@
 import 'dart:ui';
+
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+
 import 'app_color.dart';
 
 class CustomDialogBox extends StatefulWidget {
@@ -30,53 +32,54 @@ class _CustomDialogBoxState extends State<CustomDialogBox> {
     return Stack(
       children: <Widget>[
         Container(
-          padding: EdgeInsets.only(left: 20, top: 65, right: 20, bottom: 20),
-          margin: EdgeInsets.only(top: 45),
+          padding:
+              const EdgeInsets.only(left: 20, top: 65, right: 20, bottom: 20),
+          margin: const EdgeInsets.only(top: 45),
           decoration: BoxDecoration(
               shape: BoxShape.rectangle,
               color: Colors.white,
               borderRadius: BorderRadius.circular(20),
-              boxShadow: [
+              boxShadow: const [
                 BoxShadow(
                     color: Colors.black, offset: Offset(0, 10), blurRadius: 10),
               ]),
           child: Column(
             mainAxisSize: MainAxisSize.min,
             children: <Widget>[
-              Text(
+              const Text(
                 'Cnic Scanner',
                 style: TextStyle(fontSize: 18, fontWeight: FontWeight.w600),
               ),
-              SizedBox(
+              const SizedBox(
                 height: 15,
               ),
-              Text(
+              const Text(
                 'Please select any option',
                 style: TextStyle(fontSize: 15),
                 textAlign: TextAlign.center,
               ),
-              SizedBox(
+              const SizedBox(
                 height: 22,
               ),
               Row(
                 mainAxisAlignment: MainAxisAlignment.spaceAround,
                 children: [
-                  FlatButton(
+                  TextButton(
                       onPressed: () {
                         Navigator.pop(context);
                         widget.onCameraBTNPressed();
                       },
-                      child: Text(
+                      child: const Text(
                         'CAMERA',
                         style: TextStyle(
                             fontSize: 18, color: Color(kDeepDarkGreenColor)),
                       )),
-                  FlatButton(
+                  TextButton(
                       onPressed: () {
                         Navigator.pop(context);
                         widget.onGalleryBTNPressed();
                       },
-                      child: Text(
+                      child: const Text(
                         'GALLERY',
                         style: TextStyle(fontSize: 18, color: Colors.green),
                       )),
@@ -92,7 +95,7 @@ class _CustomDialogBoxState extends State<CustomDialogBox> {
             backgroundColor: Colors.transparent,
             radius: 45,
             child: ClipRRect(
-                borderRadius: BorderRadius.all(Radius.circular(45)),
+                borderRadius: const BorderRadius.all(Radius.circular(45)),
                 child: Image.asset("assets/images/person_icon.png")),
           ),
         ),
